script: move testbinding DOM interfaces to script/dom/test
