+++ components/script/dom/mod.rs
@@ -455,29 +455,9 @@ pub(crate) mod svggraphicselement;
 pub(crate) mod svgimageelement;
 pub(crate) mod svgsvgelement;
 #[cfg(feature = "testbinding")]
-pub(crate) mod testbinding;
+mod test;
 #[cfg(feature = "testbinding")]
-pub(crate) mod testbindingiterable;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingmaplikewithinterface;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingmaplikewithprimitive;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingpairiterable;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingproxy;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingsetlikewithinterface;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testbindingsetlikewithprimitive;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testns;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testutils;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testworklet;
-#[cfg(feature = "testbinding")]
-pub(crate) mod testworkletglobalscope;
+pub(crate) use self::test::*;
 pub(crate) mod text;
 pub(crate) mod textcontrol;
 pub(crate) mod textdecoder;
