+++ components/script/dom/test/mod.rs
@@ -1,0 +1,16 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod testbinding;
+pub(crate) mod testbindingiterable;
+pub(crate) mod testbindingmaplikewithinterface;
+pub(crate) mod testbindingmaplikewithprimitive;
+pub(crate) mod testbindingpairiterable;
+pub(crate) mod testbindingproxy;
+pub(crate) mod testbindingsetlikewithinterface;
+pub(crate) mod testbindingsetlikewithprimitive;
+pub(crate) mod testns;
+pub(crate) mod testutils;
+pub(crate) mod testworklet;
+pub(crate) mod testworkletglobalscope;
