+++ components/script/dom/test/testbindingproxy.rs
