+++ components/script/dom/test/testbindingsetlikewithinterface.rs
@@ -8,10 +8,10 @@ use dom_struct::dom_struct;
 use indexmap::IndexSet;
 use js::rust::HandleObject;
 
-use super::bindings::like::Setlike;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::TestBindingSetlikeWithInterfaceBinding::TestBindingSetlikeWithInterfaceMethods;
 use crate::dom::bindings::error::Fallible;
+use crate::dom::bindings::like::Setlike;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object_with_proto};
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::globalscope::GlobalScope;
