+++ components/script/dom/test/testutils.rs
@@ -9,8 +9,8 @@ use js::jsapi::{GCReason, JS_GC};
 use script_bindings::reflector::Reflector;
 use script_bindings::script_runtime::CanGc;
 
-use super::globalscope::GlobalScope;
 use crate::dom::bindings::codegen::Bindings::TestUtilsBinding::TestUtilsMethods;
+use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::test::TrustedPromise;
