+++ components/script/dom/test/testworkletglobalscope.rs
